+++ src/Blog/BlogModule.php
@@ -4,8 +4,9 @@ namespace App\Blog;
 
 use Framework\Router;
 use Psr\Container\ContainerInterface;
-use App\Blog\Controller\HomepageController;
 use Framework\Renderer\RendererInterface;
+use App\Blog\Controller\HomepageController;
+use App\Blog\Controller\PostShowController;
 
 class BlogModule
 {
@@ -16,6 +17,9 @@ class BlogModule
         //inject router
         $router = $container->get(Router::class);
         //declare route
+        //route homepage
         $router->get('/blog', HomepageController::class, 'blog.index');
+        //route post show
+        $router->get('/blog/{id:[0-9]+}', PostShowController::class, 'blog.show');
     }
 }
