+++ src/Blog/views/index.html.twig
@@ -7,7 +7,7 @@
 		<div class="row">
 			<div class="col-lg-8 col-md-10 mx-auto">
 				<div class="post-preview">
-					<a href="#">
+					<a href="{{path('blog.show',{id:item.id})}}">
 						<h2 class="post-title">
 							{{item.title}}
 						</h2>
