+++ src/Framework/App.php
@@ -44,15 +44,21 @@ class App
 
         //get route with match method of router
         $route = $router->match($request);
-
+        
         //if route is nul send error 404
         if (is_null($route)) {
             //create response 404
             return new Response(404, [], '<h1>Erreur 404</h1>');
         }
 
-         //get params from route
-         $params = $route->getParams();
+        //get params from route
+        $params = $route->getParams();
+        
+        //send atttibute to request
+        $request = array_reduce(array_keys($params), function ($request, $key) use ($params) {
+            return $request->withAttribute($key, $params[$key]);
+        }, $request);
+        
         //get callback from route
         $callback = $route->getCallback();
