+++ views/layout-front.html.twig
@@ -21,7 +21,7 @@
 		href='https://fonts.googleapis.com/css?family=Open+Sans:300italic,400italic,600italic,700italic,800italic,400,300,600,700,800' rel='stylesheet' type='text/css'>
 
 		<!-- Custom styles for this template -->
-		<link href="front/css/clean-blog.min.css" rel="stylesheet">
+		<link href="../front/css/clean-blog.min.css" rel="stylesheet">
 
 	</head>
 
@@ -46,14 +46,16 @@
 		</nav>
 
 		<!-- Page Header -->
-		<header class="masthead" style="background-image: url('img/front.jpg')">
+		<header class="masthead" style="background-image: url({% block background '../img/front.jpg' %})">
 			<div class="overlay"></div>
 			<div class="container">
 				<div class="row">
 					<div class="col-lg-8 col-md-10 mx-auto">
 						<div class="site-heading">
+							{% block heading %}
 							<h1>Xavier Jeanne</h1>
 							<span class="subheading">Développeur in progress</span>
+							{% endblock %}
 						</div>
 					</div>
 				</div>
@@ -107,6 +109,6 @@
 				timeago().render(document.querySelectorAll('.timeago'), 'fr');
 			</script>
 			<!-- Custom scripts for this template -->
-			<script src="front/js/clean-blog.min.js"></script>
+			<script src="../front/js/clean-blog.min.js"></script>
 		</body>
 	</html>
